refactor: Reorganize src/ by function, eliminate legacy/ folder
Reorganized the entire src/ directory structure from legacy-based to function-based
organization. The legacy/ folder has been completely removed and files redistributed
to appropriate directories based on their purpose.

File movements:
- src/legacy/scancodes.* → src/core/ (domain constants)
- src/legacy/modifiers.* → src/core/ (domain constants)
- src/legacy/led.* → src/platform/ (Windows LED API)
- src/legacy/traybar.* → src/platform/ (Windows system tray)
- src/legacy/configUtils.* → src/config/ (INI configuration)
- src/legacy/utils.* → src/utils/StringUtils.* (generic utilities)
- src/legacy/capsicain_legacy.* → src/CapsicainImpl.* (main implementation)

Directory structure after reorganization:
├── src/
│   ├── CapsicainImpl.cpp/h      # Main entry point (from legacy)
│   ├── core/                     # Domain constants & types
│   │   ├── scancodes.cpp/h
│   │   ├── modifiers.cpp/h
│   │   └── (existing Types.h, KeyEvent.h)
│   ├── config/                   # Configuration management (NEW)
│   │   └── configUtils.cpp/h
│   ├── utils/                    # Generic utilities (NEW)
│   │   └── StringUtils.cpp/h
│   ├── platform/                 # Platform-specific Windows code
│   │   ├── led.cpp/h            # (from legacy)
│   │   ├── traybar.cpp/h        # (from legacy)
│   │   └── (existing interception, pch, etc.)
│   ├── services/                # Service layer
│   ├── domain/                  # Business logic
│   ├── app/                     # Application layer
│   ├── commands/                # Command handling
│   └── ui/                      # User interface

Updated files:
- CMakeLists.txt: Updated all file paths and reorganized source lists
- tests/CMakeLists.txt: Updated test source paths
- All source files: Updated 41+ #include statements
- StringUtils.cpp: Fixed self-include path

Impact:
- src/legacy/ folder DELETED (empty after reorganization)
- Zero "legacy" references remaining in codebase
- All 196 tests passing
- Clean, logical directory structure

This completes the architectural refactoring initiated in Phase 1-5. The codebase
now has a clear, maintainable structure organized by function rather than history.
