+++ src/CapsicainImpl.h
@@ -1,0 +1,121 @@
+#include <string>
+#include <map>
+#include <cstdint>
+
+#include "platform/interception.h"
+#include "utils/StringUtils.h"
+#include "config/configUtils.h"
+#include "platform/traybar.h"
+
+// Forward declaration for RuntimeStateService
+namespace capsicain { namespace services { class RuntimeStateService; } }
+
+// Debug logging - runtime conditional based on options
+#define IFDEBUG if(options.debug && !globalState.secretSequenceRecording)
+
+// Trace and profiling - compile-time disabled (can be re-enabled by changing to true)
+constexpr bool ENABLE_TRACE = false;
+constexpr bool ENABLE_PROFILING = false;
+
+struct Executable
+{
+    std::string verb;
+    std::string path;
+    std::string args;
+    std::string dir;
+    int mode;
+    HANDLE proc;
+    DWORD pid;
+    HWND hwnd;
+};
+
+struct Device {
+    std::string id;
+    bool keyboard;
+    bool apple;
+};
+
+enum KEYSTATE
+{
+    KEYSTATE_DOWN = 0,
+    KEYSTATE_UP = 1,
+    KEYSTATE_E0_DOWN = 2,
+    KEYSTATE_E0_UP = 3,
+    KEYSTATE_E1_DOWN = 4,
+    KEYSTATE_E1_UP = 5
+};
+
+void keySequenceAppendMakeKey(unsigned short scancode, std::vector<VKeyEvent> &sequence);
+void keySequenceAppendBreakKey(unsigned short scancode, std::vector<VKeyEvent> &sequence);
+void keySequenceAppendMakeBreakKey(unsigned short scancode, std::vector<VKeyEvent> &sequence);
+
+std::string getSymbolForIKStrokeState(unsigned short state);
+
+bool processOnOffKey(capsicain::services::RuntimeStateService& runtimeState, uint8_t scancode, bool isDownstroke);
+void InterceptionSendCurrentKeystroke();
+void handleConfigSwitch(int scancode);
+
+// ============================================================================
+// REMOVED: Unused legacy function declarations (Phase 5 cleanup)
+// ============================================================================
+// The following functions were declaration-only (no implementation) and have
+// been completely replaced by domain classes:
+//
+// - processModifierState() → ModifierTracker class
+// - detectTapping() → TapDetector class
+// - processRewireScancodeToVirtualcode() → KeyMapper class
+// - processCombos() → ComboMatcher class
+// - processMapAlphaKeys() → KeyMapper class
+//
+// Note: processMessyKeys() implementation still exists but is unused
+// ============================================================================
+
+void playKeyEventSequence(std::vector<VKeyEvent> keyEventSequence);
+
+void printOptions();
+
+void sendVKeyEvent(VKeyEvent keyEvent, bool hold = true);
+
+// REMOVED: sendResultingKeyOrSequence() - Replaced by KeyProcessingService
+
+VKeyEvent convertIkstroke2VKeyEvent(InterceptionKeyStroke ikStroke);
+
+void normalizeIKStroke(InterceptionKeyStroke &ikstroke);
+InterceptionKeyStroke convertVkeyEvent2ikstroke(VKeyEvent keyEvent);
+std::map<uint8_t, Device>* getHardwareId(bool refresh = true);
+
+bool initConsoleWindow();
+
+// Forward declaration for ConfigurationService
+namespace capsicain { namespace services { class ConfigurationService; } }
+void parseIniGlobals(capsicain::services::ConfigurationService& configService);
+
+void printHelloHeader();
+void printStatus();
+void printKeylabels();
+void printHelp();
+void printIKStrokeState(InterceptionKeyStroke iks);
+void printLoopState1Input();
+void printLoopState2Modifier();
+void printLoopStateMappingTime(long us);
+void printLoopState4TapState();
+
+
+void reset();
+void reload(capsicain::services::ConfigurationService& configService);
+void releaseAllSentKeys();
+std::vector<std::string> assembleConfig(int config);
+void switchConfig(int config, bool forceReloadSameLayer);
+void resetCapsNumScrollLock();
+
+int obfuscateVKey(int vk);
+int deObfuscateVKey(int vk);
+
+int getKeyHolding(int vcode);
+bool runExecutable(Executable &exe);
+void killExecutable(Executable &exe);
+void killExecutableByPath(std::string path);
+
+void loadAHK();
+void unloadAHK();
+void sendAHK(const std::string& msg);
