+++ src/platform/constants.h
