+++ src/platform/resource.h
