+++ src/services/KeyProcessingService.cpp
@@ -3,8 +3,8 @@
 #include "services/RuntimeStateService.h"
 #include "services/HardwareService.h"
 #include "services/UIService.h"
-#include "legacy/scancodes.h"
-#include "legacy/modifiers.h"
+#include "core/scancodes.h"
+#include "core/modifiers.h"
 
 namespace capsicain {
 namespace services {
